Update DrawingPanel.java
The whole crane is there in parts but i'm going to change the bucket so its kind of just floating around. no picking, etc. also I lost her original translation code for when the whole crane is being moved, i have the "if base or platform selected == allobject" but the code within that isn't correct.